Initialized separate cylinder components with correct offests. TODO Linear interpolate offsets per frame.

+++ Source/JediTrainer/Public/CylinderComponent.h
@@ -30,9 +30,9 @@ public:
 	// Set the scale of the cylinder transform
 	// TODO Would it be preferable to take in a vector?
 	UFUNCTION(BlueprintCallable, Category = "Switch Functions") void SetScale(
-		UPARAM(DisplayName = "Width") int w, // Corresponds to x axis
-		UPARAM(DisplayName = "Depth") int d, // Corresponds to y axis
-		UPARAM(DisplayName = "Height") int h // Corresponds to z axis
+		UPARAM(DisplayName = "Width") float w, // Corresponds to x axis
+		UPARAM(DisplayName = "Depth") float d, // Corresponds to y axis
+		UPARAM(DisplayName = "Height") float h // Corresponds to z axis
 		);
 
 	// Sets the world transform of the cylinder to the given transform. Ignores parenting.
